Basic Implementation Problem.
AC code on contest Time.

NB: I solved this problem in one go but for a typo (I wrote "web" instead of "wed" - Wednesday) and got 8 wrong answer verdicts for that. That's why I was not selected for the onsite contest. While I am uploading this solution it reminds me of the moment. 😢😢